07-AI-Chatbot/requirements.txt
livekit-agents[groq,silero,turn-detector]~=1.0rc
python-dotenv